app/requirements.txt
google-cloud-storage==1.41.0
pymongo>=4.4